Relocate Command interactive completers from `org.grails.cli.interactive.completers` to `org.grails.cli.command.completers`
