+++ grace-shell/src/main/groovy/org/grails/cli/GrailsCli.groovy
@@ -56,14 +56,14 @@ import org.grails.cli.command.CommandCompleter
 import org.grails.cli.command.CommandRegistry
 import org.grails.cli.command.ExecutionContext
 import org.grails.cli.command.ProjectContext
+import org.grails.cli.command.completers.EscapingFileNameCompletor
+import org.grails.cli.command.completers.RegexCompletor
+import org.grails.cli.command.completers.SortedAggregateCompleter
+import org.grails.cli.command.completers.StringsCompleter
 import org.grails.cli.compiler.dependencies.GrailsDependenciesDependencyManagement
 import org.grails.cli.gradle.ClasspathBuildAction
 import org.grails.cli.gradle.GradleAsyncInvoker
 import org.grails.cli.gradle.cache.MapReadingCachedGradleOperation
-import org.grails.cli.interactive.completers.EscapingFileNameCompletor
-import org.grails.cli.interactive.completers.RegexCompletor
-import org.grails.cli.interactive.completers.SortedAggregateCompleter
-import org.grails.cli.interactive.completers.StringsCompleter
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.ProfileRepository
 import org.grails.cli.profile.repository.GrailsRepositoryConfiguration
