+++ grace-shell/src/main/groovy/org/grails/cli/command/completers/ClosureCompleter.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2025 the original author or authors.
+ * Copyright 2014-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,32 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.interactive.completers
+package org.grails.cli.command.completers
 
 import groovy.transform.CompileStatic
-import org.springframework.core.io.Resource
-
-import grails.util.BuildSettings
-
-import org.grails.io.support.GrailsResourceUtils
+import jline.console.completer.Completer
 
 /**
- * A completer for domain classes
- *
  * @author Graeme Rocher
- * @author Michael Yan
  * @since 3.0
  */
 @CompileStatic
-class DomainClassCompleter extends ClassNameCompleter {
+class ClosureCompleter implements Completer {
 
-    DomainClassCompleter() {
-        super(new File(BuildSettings.GRAILS_APP_DIR, 'domain'))
+    private final Closure<List<String>> closure
+    private Completer completer
+
+    ClosureCompleter(Closure<List<String>> closure) {
+        this.closure = closure
+    }
+
+    Completer getCompleter() {
+        if (completer == null) {
+            completer = new jline.console.completer.StringsCompleter(closure.call())
+        }
+        completer
     }
 
     @Override
-    boolean isValidResource(Resource resource) {
-        GrailsResourceUtils.isDomainClass(resource.getURL())
+    int complete(String buffer, int cursor, List<CharSequence> candidates) {
+        getCompleter().complete(buffer, cursor, candidates)
     }
 
 }
