+++ grace-shell/src/main/groovy/org/grails/cli/command/completers/EscapingFileNameCompletor.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2022 the original author or authors..
+ * Copyright 2011-2026 the original author or authors..
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.interactive.completers
+package org.grails.cli.command.completers
 
 import jline.console.completer.FileNameCompleter
