+++ grace-shell/src/main/groovy/org/grails/cli/core/template/TemplateRendererImpl.groovy
@@ -28,9 +28,9 @@ import org.springframework.core.io.ResourceLoader
 import grails.codegen.model.Model
 
 import org.grails.cli.command.ExecutionContext
+import org.grails.cli.command.completers.ClassNameCompleter
 import org.grails.cli.core.io.FileSystemInteraction
 import org.grails.cli.core.io.FileSystemInteractionImpl
-import org.grails.cli.interactive.completers.ClassNameCompleter
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.ProfileRepository
 import org.grails.cli.profile.ProfileRepositoryAware
