+++ grace-shell/src/main/groovy/org/grails/cli/profile/steps/RenderStep.groovy
@@ -24,8 +24,8 @@ import grails.util.GrailsNameUtils
 
 import org.grails.build.parsing.CommandLine
 import org.grails.cli.command.ExecutionContext
+import org.grails.cli.command.completers.ClassNameCompleter
 import org.grails.cli.core.template.SimpleTemplate
-import org.grails.cli.interactive.completers.ClassNameCompleter
 import org.grails.cli.profile.AbstractStep
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.support.ArtefactVariableResolver
